+++ packages/e2e-tests/scripts/deployAndDeposit.ts
@@ -11,17 +11,20 @@ import {
   CanonAddress,
   DepositRequest,
   StealthAddressTrait,
+  zip,
 } from "@nocturne-xyz/sdk";
 import { KEYS_TO_WALLETS } from "../src/keys";
 import { SimpleERC20Token } from "@nocturne-xyz/contracts/dist/src/SimpleERC20Token";
 
-const HH_URL = "http://localhost:8545";
+const ANVIL_URL = "http://127.0.0.1:8545";
 
-// Corresponds to privkey 0x000...005
-const SUBTREE_BATCH_FILLER = "0xe1AB8145F7E55DC933d51a18c793F901A3A0b276";
+// anvil account #5
+const SUBTREE_BATCH_FILLER = "0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc";
 
-// Corresponds to privkey 0x000...006
-const DEPOSIT_SCREENER = "0xE57bFE9F44b819898F47BF37E5AF72a0783e1141";
+// anvil account #6
+const DEPOSIT_SCREENER = "0x976EA74026E726554dB657fA54763abd0C3a0aa9";
+const TEN_GWEI= 10n * 10n ** 9n;
+const GAS_TOKEN_AMOUNT = 10000000n ** 9n * TEN_GWEI;
 
 // add MM Flask addresses here
 const TEST_ETH_ADDRS = [
@@ -47,7 +50,7 @@ const TEST_CANONICAL_NOCTURNE_ADDRS: CanonAddress[] = [
 
 (async () => {
   console.log("deploying contracts with dummy proxy admin...");
-  const provider = new ethers.providers.JsonRpcProvider(HH_URL);
+  const provider = new ethers.providers.JsonRpcProvider(ANVIL_URL);
   const chainId = BigInt((await provider.getNetwork()).chainId);
   const [deployerEoa] = KEYS_TO_WALLETS(provider);
 
@@ -62,34 +65,49 @@ const TEST_CANONICAL_NOCTURNE_ADDRS: CanonAddress[] = [
     deployerEoa
   );
 
+  const tokenAmount = ethers.utils.parseEther("10.0").toBigInt();
   const tokenFactory = new SimpleERC20Token__factory(deployerEoa);
   const tokens: SimpleERC20Token[] = [];
+  const amounts: bigint[] = [];
   for (let i = 0; i < 2; i++) {
     const token = await tokenFactory.deploy();
     await token.deployed();
     console.log(`Token ${i + 1} deployed at: ${token.address}`);
     tokens.push(token);
+    amounts.push(tokenAmount);
   }
 
-  // airdrop ETH and reserve test tokens (outside nocturne) to each addr in `TEST_ETH_ADDRS`
-  for (const token of tokens) {
+  for (const [token, amount] of zip(tokens, amounts)) {
+    // airdrop ETH and reserve test tokens (outside nocturne) to each addr in `TEST_ETH_ADDRS`
     for (const addr of TEST_ETH_ADDRS) {
       console.log(`Sending ETH and tokens to ${addr}`);
-      await deployerEoa.sendTransaction({
-        to: addr,
-        value: ethers.utils.parseEther("10.0"),
-      });
-      await token.reserveTokens(addr, ethers.utils.parseEther("10.0"));
+      {
+        const tx = await deployerEoa.sendTransaction({
+          to: addr,
+          value: ethers.utils.parseEther("10.0"),
+        });
+        await tx.wait(1);
+      } 
+      {
+        const tx = await token.reserveTokens(addr, amount);
+        await tx.wait(1);
+      }
     }
 
     // Reserve and approve tokens for nocturne addr deployer
-    const reserveAmount = ethers.utils.parseEther("100.0");
-    await token
-      .connect(deployerEoa)
-      .reserveTokens(deployerEoa.address, reserveAmount);
-    await token
-      .connect(deployerEoa)
-      .approve(depositManager.address, reserveAmount);
+    const reserveAmount = amount * 100n;
+    {
+      const tx = await token
+        .connect(deployerEoa)
+        .reserveTokens(deployerEoa.address, reserveAmount);
+      await tx.wait(1);
+    }
+    {
+      const tx = await token
+        .connect(deployerEoa)
+        .approve(depositManager.address, reserveAmount);
+      await tx.wait(1);
+    }
   }
 
   const encodedAssets = tokens
@@ -104,29 +122,29 @@ const TEST_CANONICAL_NOCTURNE_ADDRS: CanonAddress[] = [
   const targetAddrs = TEST_CANONICAL_NOCTURNE_ADDRS.map(
     StealthAddressTrait.fromCanonAddress
   );
-  const perNoteAmount = ethers.utils.parseEther("10.0");
-  for (const encodedAsset of encodedAssets) {
+  for (const [encodedAsset, amount] of zip(encodedAssets, amounts)) {
     // Deposit two 100 unit notes for given token
     for (const addr of targetAddrs) {
-      console.log("depositing 1 100 token note to", addr);
+      console.log(`depositing 1 ${amount} note to`, addr, `from ${deployerEoa.address}`);
 
       const nonce = await depositManager._nonces(deployerEoa.address);
       const depositRequest: DepositRequest = {
         chainId,
         spender: deployerEoa.address,
         encodedAsset,
-        value: perNoteAmount.toBigInt(),
+        value: amount,
         depositAddr: addr,
         nonce: nonce.toBigInt(),
         gasCompensation: BigInt(0),
       };
 
-      const instantiateDepositTx = await depositManager
+      const tx = await depositManager
         .connect(deployerEoa)
         .instantiateDeposit(depositRequest);
-      await instantiateDepositTx.wait(1);
+      await tx.wait(1);
     }
   }
 
-  await handler.connect(deployerEoa).fillBatchWithZeros();
+  const tx = await handler.connect(deployerEoa).fillBatchWithZeros();
+  await tx.wait(1);
 })();