+++ packages/subgraph/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Unreleased
 
+- Update deposit manager address
 - Remove `wallet` in place of `handler` contract for syncing sdk events
 - Add entities/events/handler for DepositManager `DepositInstantiated` event